community/requirements.txt
amqp==2.1.4
appdirs==1.4.3
asgi-ipc==1.3.1
asgi-redis==1.3.0
asgiref==1.1.1
asn1crypto==0.22.0
attrs==16.3.0
autobahn==17.5.1
Automat==0.5.0
backports-abc==0.5
billiard==3.5.0.2
celery==4.0.2
certifi==2017.7.27.1
cffi==1.10.0
channels==1.1.3
chardet==3.0.4
chaussette==1.3.0
constantly==15.1.0
cryptography==1.8.1
Cython==0.25.2
daphne==1.2.0
decorator==4.0.11
Django==1.11
django-autoslug==1.9.3
django-bootstrap-form==3.2.1
django-debug-toolbar==1.7
django-debug-toolbar-line-profiler==0.6.1
django-debug-toolbar-template-timings==0.7
django-guardian==1.4.8
django-ordered-model==1.4.1
django-postgresql-manager==0.2.1
django-redis-cache==1.7.1
django-simple-history==1.8.2
django-steamauth==1.1.0
django-uwsgi==0.2.0
docopt==0.6.2
gevent==1.2.1
gevent-eventemitter==2.0
greenlet==0.4.12
hiredis==0.2.0
idna==2.6
incremental==16.10.1
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==4.0.2
line-profiler==2.0
monotonic==1.3
msgpack-python==0.4.8
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
posix-ipc==1.0.0
prompt-toolkit==1.0.14
protobuf==3.3.0
psycopg2==2.7.1
ptyprocess==0.5.1
pycparser==2.17
pycrypto==2.6.1
Pygments==2.2.0
Pympler==0.5
pyparsing==2.2.0
python-redis-lock==3.2.0
python-valve==1.0.0
pytz==2017.2
redis==2.10.6
redis-structures==0.1.6
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
skcode==3.2.1
sqlparse==0.2.3
steam==0.8.18
steamodd==4.21
tornado==4.5.2
traitlets==4.3.2
Twisted==17.1.0
txaio==2.7.1
typing==3.6.1
urllib3==1.22
uWSGI==2.0.15
vdf==2.3
vine==1.1.3
vital-tools==0.1.7
wcwidth==0.1.7
zope.interface==4.4.0
